requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4==4.12.2
pydantic>=2.11.9
scrapy==2.11.0
lxml>=5.0.0
requests>=2.28.0
urllib3==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0